MODFQMMGR-79 Add filterValueGetter to entityTypeColumn
This new property will allow us to specify a column accessor that is
different from what the user actually sees. The current use case for
this is columns with an index on left(lower(f_unaccent(column)), 600).
We want to use that form in the WHERE clause of queries, but want the
user to see it without the extra processing.