MODFQMMGR-378 Update approach to enabling ECS for an entity type
This commit adds a new boolean crossTenantQueriesEnabled property to
entity types, to control when an entity type supports cross-query
queries. It also changes the shape of the entity type column "source"
property's object model, adding a new "type" property. This new property
will allow us to use the "source" to define value sources that come from
things besides entity types. There's also a new "name" property, which
is for setting the name of the value provider within the source type