FQTM-12 Update ET model to support custom ETs
This commit adds a subclass of the EntityType class, which adds some
extra fields that will only exist in custom entity types. It also
adds a description field to the base EntityType class.

In addition to those changes, it also adds a new name property to the ET
source class, which will override source name translations when set.
This is to support user-friendly source/field names in custom entity
types